app/requirements.txt
fastapi[standard]
SQLAlchemy
alembic
pydantic-settings
passlib
bcrypt
python-jose
PyGithub
kubernetes
pymysql